Add support for symbol renaming
